feat(biologic): select XY
Still working on it. This will select XY variables that are useful for
the user depending on the technique set out in Electro Chemical Lab
software, and carried out by the BioLogic hardware.

The code is more or less complete for it but still needs good defaults
for about 5 other techniques.