fix(autoPicking): use current value of T
Explain my idea in case it is wrong:

* It seems the calculation measures how much is absorbed. A high
value of T gives `w` as this is passing through the cell.

I changed it from array to number.

Imho the Intensity value should be just "s" not "S". I didn't change it in case it
was in purpose for some reason that I can't get.