chore: flag usage of implicit any (#397)
* chore(tsconfig.json): flag usage of implicit any

* refactor: separate File[] from FileCollection

Index error came from `FileCollection[i]`. I tried to avoid duplicating
stuff (code and references to arrays).

Also updated a dependency to include missing types

* rollback(babel types): remove unnecesary type

* test(fix): error pick type coming from IR autoPick

* fix(autoPicking): use current value of T

Explain my idea in case it is wrong:

* It seems the calculation measures how much is absorbed. A high
value of T gives `w` as this is passing through the cell.

I changed it from array to number.

Imho the Intensity value should be just "s" not "S". I didn't change it in case it
was in purpose for some reason that I can't get.

* refactor(enhancers): add type, remove anys

* fix(type): pass the enhancer to loadMeasurements

* chore(types): pull variable off the for in loop

* fix(types): orientation and error type

* fix(type): netcdf reader and MExplorer

* fix(loader): remove explicit any

* feat(biologic): add OCV and fix types

* fix(package): update loaders

* fix(tsconfig): remove allowJS

* chore(tsconfig): use strict only

* update deps

* fix

* update build target

Co-authored-by: Michaël Zasso <targos@protonmail.com>