refactor: redo toolbar with bluprintjs components
BREAKING CHANGE: the api of the `Toolbar` component has changed. To migrate, set `minimal=true` prop and if `orientation` was defined as `vertical`, set `vertical=true` prop. The api of the `Toolbar.Item` has changed. To migrate, set the icon in the `icon` prop instead of passing it in children.