refactor!: migrate Toolbar to blueprintjs components (#559)
BREAKING CHANGE: the api of the `Toolbar` component has changed. To migrate, remove `orientation='horizontal'` or replace `orientation='vertical'` with `vertical={true}`.  The api of the `Toolbar.Item` has changed. To migrate, set the icon in the `icon` prop instead of passing it in children. It is recommended to use blueprint icons by passing one of the available string constants to `icon`.