feat!: improve Toolbar component
BREAKING CHANGE: Rename `title` to `tooltip` and set it as optional, and it accepts string or JSX.Element.
BREAKING CHANGE: Remove `noTooltip` since the tooltip is an optional property.
feat: Add an optional property `tooltipProps` which allows to customize the appearance of the tooltip
refactor: use the the new Toolbar API.