feat(SplitPane): rework SplitPane API
Refactor the SplitPane API to be more consistent and idiomatic.
Make the `open` and `size` state of the component controllable.

BREAKING CHANGE: The `onToggle` prop has been renamed to `onOpenChange`
and receives true when the split pane opens (previously received false). 
Contrary to the previous onToggle, onOpenChange is called in any
circumstance when the split pane state changes, for example also when it
automatically closes after it reaches the `closeThreshold` size.
The original behavior of the `closed` and `size` props has changed.
Their behavior were a mix of a control value prop and default value prop
for internally managed state.
There is now a clear separation between control props (`open` and `size`)
and default value props `defaultOpen` and `defaultSize`.
For example for the open state you should either:
- Use `open` to control if the SplitPane is open or not, and `onOpenChange` to react to changes. Using `defaultOpen` will have no effect.
- Use `defaultOpen` to set the initial value (only affects first render), with state changes being handled internally by the component.
If you used `closed` with a number, use the new `closeThreshold` prop instead.


Closes: https://github.com/zakodium-oss/react-science/issues/852