feat(Accordion): allow state to be controlled and rework API.
Since the title acts as an identifier, allow to constrain its type to a literal union.

Allow to choose between controlled and uncontrolled for AccordionItem's open state.

Add `closeAllExcept` method to the useToggleAccordion hook.

Print an error to the console when a title is reused.

Closes: https://github.com/zakodium-oss/react-science/issues/876

BREAKING-CHANGE: the `useAccordionContext` hook has been removed from the API. It should only be used internally. The `defaultOpened` prop of `Accordion.Item` has been renamed to `defaultOpen`.