feat: change FullScreenProvider API to not create additional div (#897)
BREAKING-CHANGE: FullScreenProvider now takes a render callback with the ref to attach to the element which can be displayed fullscreen. It no longer renders an additional styled div.
The RootLayout is now a fullscreenable element. It is no longer needed to wrap it in a FullScreenProvider around it.
A global CSS change makes the backdrop of fullscreen elements white.

Closes: https://github.com/zakodium-oss/react-science/issues/689