feat: improve fullscreen error handling
Use a library to manage fullscreen APIs.
Also add a `onToggleError` prop to the `FullScreenProvider` and `RootLayout` components.

There is a behavior change with "stacked" fullscreen elements (like in the storybook).
Now when you enter fullscreen a second time, it will first exit the previous fullscreen.
I think it's fine. We don't (and probably shouldn't) use stacked fullscreens in real apps.