feat(Accordion): improve renderToolbar API (#902)
Ref: https://github.com/zakodium/peak-by-peak/issues/360

BREAKING-CHANGE: The render callback parameter has changed to contain multiple properties instead of just the `isOpen` state. It can now be used to control the state of the item in which it is rendered. Click events in the toolbar no longer propagate to the header in order to prevent it from toggling if it is not wanted.